requirements.txt
# Create requirements.txt
requirements_content = """
# Data manipulation and analysis
pandas>=1.3.0
numpy>=1.21.0
scipy>=1.7.0

# Machine Learning
scikit-learn>=1.0.0
xgboost>=1.4.0
lightgbm>=3.2.0

# Deep Learning
tensorflow>=2.6.0
keras>=2.6.0
torch>=1.9.0
torchvision>=0.10.0

# Time Series Analysis
statsmodels>=0.12.0
pmdarima>=1.8.0

# Data Visualization
matplotlib>=3.4.0
seaborn>=0.11.0
plotly>=5.0.0

# Utilities
tqdm>=4.62.0
joblib>=1.0.0
pickle-mixin>=1.0.2

# Data fetching
requests>=2.25.0
ucimlrepo>=0.0.3

# Configuration
pyyaml>=5.4.0
configparser>=5.0.0

# Logging
loguru>=0.5.0

# Jupyter notebook support
jupyter>=1.0.0
ipykernel>=6.0.0
""".strip()

with open('household_energy_consumption_ml/requirements.txt', 'w') as f:
    f.write(requirements_content)

print("requirements.txt created successfully!")